feat(order): implement state machine, cancellation logic and admin statistics
- Refactor `updateOrderStatus` with a strict state transition matrix and RBAC.
- Implement `cancelOrder` for active users and `@Scheduled` `cancelUnpaidOrders` for system-level cleanup.
- Add administrative statistics: `countUniqueCustomers` and `getRevenueByDateRange`.
- Ensure defense-in-depth with Service-layer authorization and read-only transactions.
- Fix: Use BigDecimal.valueOf(quantity) for accurate subtotal calculation.

+++ src/main/java/io/github/j_yuhanwang/food_ordering_app/order/services/OrderService.java
@@ -29,13 +29,13 @@ public interface OrderService {
 
     //3.change the status
     OrderDTO updateOrderStatus(Long orderId,OrderStatus status);
-    //Timed scanning method(for 5 minutes, do not convey to frontend)
+    //Timed scanning method (waiting for 15 minutes, do not convey to frontend)
     void cancelUnpaidOrders();
-    //user cancel the order
+    //user cancel the order actively
     void cancelOrder(Long orderId);
 
     //4. the aggregate information
-    int countUniqueCustomers();
+    long countUniqueCustomers();
     BigDecimal getRevenueByDateRange(LocalDateTime startDate, LocalDateTime endDate);
 
 }
